requirements.txt
Django>=5.0.7
djangorestframework>=3.16.0
django-cors-headers>=4.7.0
drf-yasg>=1.21.10
Pillow>=10.4.0
django-filter>=25.1
